添加纹理类，PBR默认材质，修复HDR过亮问题，修复多通道贴图问题

+++ OpenGL/core/Camera.cpp
@@ -28,7 +28,7 @@ glm::mat4 Camera::getViewMatrix() const
 
 glm::mat4 Camera::getProjectionMatrix() const
 {
-    return glm::perspective(glm::radians(Zoom), static_cast<float>(width) / height, Near, Far);
+    return glm::perspective(glm::radians(Zoom), static_cast<float>(width) / static_cast<float>(height), Near, Far);
 }
 
 void Camera::ProcessKeyboard(Camera_Movement direction, float deltaTime)
